command line tester application, fixed disappearing PV

+++ engine/src/main/kotlin/pl/tomaszstankowski/fourplayerchess/engine/HistoryTable.kt
@@ -7,7 +7,7 @@ internal class HistoryTable {
         }
     }
 
-    fun increase(move: MoveBits, color: Color, depth: Byte) {
+    fun increase(move: MoveBits, color: Color, depth: Int) {
         table[color.ordinal][move.from][move.to] += depth * depth
     }
